Update to queries, model names, properties and more
Changed from categories to pages for MyiWorlds. Added more properties as bases for every page.

Changed queries being on type -> path (url) -> id .  This is because it combines the id with a Datastore key and if you were to change the url of the page it would be hard.  So the path will just be indexed and still queryable.

Unable to get page: pages:[PageType] working, tried so many different things.