Adding connections, started to add AccountType for future use
Adjusted some seed files that had errors in them.

Fixed issue with global id on list making it so it was rendering the same item in a list over and over by telling it to use _id inside relay global id

Added connections (connection definition only working in same file and currently unable to seperate it into a seperate file

Still need to correctly implement other types on page.