Large update to creating Circles, and other changes
Large refactor of createCircle component.  Added snackbar for entering type, default url generation based on username, if no path is specified it creates one with the viewer who is creating the circle + uuidv1 to provide a unique id.

Added pathFull and pathName to circles.  This will be necessary for when other people are editing a circle that is not their own.  You don't want it always to be changing the root of the url to a new viewer if someone wants to edit the title of a circle.

TODOs:
Creating Circle pathFulls need to be checked in db before submitted (only need to be unique per viewer account).
When creating circle the pathName needs to be combined with viewer id on backend to prevent any hackery on frontend.